# Build: Speed up CI checks