--- a/.circleci/config.yml
+++ b/.circleci/config.yml
@@ -11,6 +11,8 @@ executors:
     working_directory: /tmp/storybook
     docker:
       - image: circleci/node:10-browsers
+        environment:
+          NODE_OPTIONS: --max_old_space_size=4096
     resource_class: <<parameters.class>>
 
 jobs: