--- a/.circleci/config.yml
+++ b/.circleci/config.yml
@@ -18,7 +18,7 @@ executors:
 jobs:
   build:
     executor:
-      class: xlarge
+      class: large
       name: sb_node
     steps:
       - add_ssh_keys:
@@ -46,7 +46,7 @@ jobs:
             - lib
   bootstrap-examples:
     executor:
-      class: medium
+      class: large
       name: sb_node
     steps:
       - add_ssh_keys:
@@ -100,7 +100,9 @@ jobs:
             cd examples/official-storybook
             yarn packtracker
   examples:
-    executor: sb_node
+    executor:
+      class: large
+      name: sb_node
     parallelism: 11
     steps:
       - add_ssh_keys:
@@ -243,7 +245,7 @@ jobs:
 
   smoke-tests:
     executor:
-      class: small
+      class: medium
       name: sb_node
     steps:
       - add_ssh_keys: