--- a/.teamcity/settings.kts
+++ b/.teamcity/settings.kts
@@ -129,8 +129,7 @@ object Build : BuildType({
                 #!/bin/bash
                 set -e -x
                 
-                yarn install
-                yarn repo-dirty-check
+                yarn install --frozen-lockfile
                 yarn bootstrap --core
             """.trimIndent()
             dockerImage = "node:10"
@@ -177,7 +176,7 @@ object ExamplesTemplate : Template({
             scriptContent = """
                 #!/bin/bash
                 set -e -x
-
+                
                 yarn install
                 rm -rf built-storybooks
                 mkdir -p built-storybooks
@@ -319,7 +318,7 @@ object E2E : BuildType({
             scriptContent = """
                 #!/bin/bash
                 set -e -x
-
+                
                 yarn install
                 yarn cypress install
                 yarn serve-storybooks &
@@ -366,42 +365,42 @@ object SmokeTests : BuildType({
             scriptContent = """
                 #!/bin/bash
                 set -e -x
-
+                
                 yarn install
-
+                
                 cd examples/cra-kitchen-sink
                 yarn storybook --smoke-test --quiet
-          
+                
                 cd ../cra-ts-kitchen-sink
                 yarn storybook --smoke-test --quiet
-          
+                
                 cd ../vue-kitchen-sink
                 yarn storybook --smoke-test --quiet
-          
+                
                 cd ../svelte-kitchen-sink
                 yarn storybook --smoke-test --quiet
-          
+                
                 cd ../angular-cli
                 yarn storybook --smoke-test --quiet
-          
+                
                 cd ../ember-cli
                 yarn storybook --smoke-test --quiet
-          
+                
                 cd ../marko-cli
                 yarn storybook --smoke-test --quiet
-          
+                
                 cd ../official-storybook
                 yarn storybook --smoke-test --quiet
-          
+                
                 cd ../mithril-kitchen-sink
                 yarn storybook --smoke-test --quiet
-          
+                
                 cd ../riot-kitchen-sink
                 yarn storybook --smoke-test --quiet
-          
+                
                 cd ../preact-kitchen-sink
                 yarn storybook --smoke-test --quiet
-          
+                
                 cd ../cra-react15
                 yarn storybook --smoke-test --quiet
             """.trimIndent()
@@ -547,7 +546,7 @@ object Test : BuildType({
                 yarn init -y
                 yarn add -D jest-teamcity
                 cd ..
-                
+
                 yarn jest --coverage -w 2 --reporters=${'$'}PWD/temp-jest-teamcity/node_modules/jest-teamcity
             """.trimIndent()
             dockerImage = "node:10"