--- a/scripts/run-registry.ts
+++ b/scripts/run-registry.ts
@@ -10,6 +10,7 @@ import nodeCleanup from 'node-cleanup';
 
 import startVerdaccioServer from 'verdaccio';
 import pLimit from 'p-limit';
+import * as os from 'os';
 import { listOfPackages, Package } from './utils/list-packages';
 
 program
@@ -106,7 +107,9 @@ const currentVersion = async () => {
 };
 
 const publish = (packages: { name: string; location: string }[], url: string) => {
-  const limit = pLimit(3);
+  const numOfCpus = os.cpus()?.length;
+
+  const limit = pLimit((numOfCpus ?? 4) - 1);
 
   return Promise.all(
     packages.map(({ name, location }) =>