--- a/scripts/utils/concurrency.js
+++ b/scripts/utils/concurrency.js
@@ -7,6 +7,21 @@ const os = require('os');
  */
 const maxConcurrentTasks = Math.max(1, os.cpus().length - 1);
 
+/**
+ * Use a simple round robin to filter input data according to the CI node currently running the script
+ * @param {Array} arrayOfData An array of anything you want
+ * @returns {Array} An array containing only the data that shoud be used by current CI node.
+ */
+function filterDataForCurrentCircleCINode(arrayOfData) {
+  const nodeIndex = +process.env.CIRCLE_NODE_INDEX || 0;
+  const numberOfNodes = +process.env.CIRCLE_NODE_TOTAL || 1;
+
+  return arrayOfData.filter((_, index) => {
+    return index % numberOfNodes === nodeIndex;
+  });
+}
+
 module.exports = {
   maxConcurrentTasks,
+  filterDataForCurrentCircleCINode,
 };