Issue: CI is taking sooooo long to run the whole CircleCI workflow. Let's try to improve that!

## What I did

 - Replace handmade `repo-dirty-check` by `--frozen-lockfile` option when running `yarn install`: 1s -> 0s
 - ~Replace CircleCI `checkout` step by a custom one to only checkout 1 commit and avoid downloading the whole monorepo and its history: 20s -> 1s for each step~ -> Not working with PR from forks see https://github.com/storybookjs/storybook/pull/12315#discussion_r500537123
 - Tune executor usages to reduce time spent on bottleneck steps (with Large or XLarge executors) and use small executors for noncritical steps: not making a big diff alone 
 - Use the number of vcore minus 1 as concurrency level in `publish-to-verdaccio` script: ~3 min -> ~2min with a large executor 
 - Rework the steps: 
    - `build`: It's a merge of previous `install` and `build` steps as there is no real benefit to have 2 different steps and it avoids to persist and restore the workspace between theses steps. It _doesn't install dependencies from examples projects_ (i.e. delete `examples` directory before yarn install) to ensure the build can be done with deps listed in `@storybook` packages and not by chance thanks to examples deps. Built packages are then available in the workspace.
    - `install-examples-deps`: Install all dependencies required by `examples` and make them available in the workspace.
    - `install-e2e-deps`: Install only dependencies needed to run e2e tests and make them available in the workspace i.e. delete all sub-directory and so do a `yarn install` with only the root `package.json`. It has at least 2 benefits: fewer dependencies so faster bootstrap of e2e tests and, ensuring we have the required deps in root `package.json`.
    - All others jobs are still the same, only the previous jobs requirement has changed. All jobs are using `build` workspace and according to their need they are using either the `node_modules` from `install-examples-deps` or from `install-e2e-deps`

![image](https://user-images.githubusercontent.com/4112568/93671500-d9dd6580-faa3-11ea-9a19-b52c31eb8c85.png)

 
Everything combined workflow is running in ~15min (according to global workload) instead of ~25min 🎉 .

Let me know if you want me to add some doc somewhere or if some parts are not clear enough.


## How to test

- CircleCI and TeamCity should still be 🟢