@gaetanmaisse 2020-09-23
I guess all TeamCity jobs are failing because it somehow uses the workflow from `next` branch and the one of this branch 🤔