@gaetanmaisse 2020-09-23
> UI Tests Pending — 41 changes must be accepted as baselines.

Not sure about chromatic check, all stories reported as changes are the same as the one from next 🤔