@ndelangen 2020-09-28
The tech/speed-up-ci-checks branch had 1 commit, and no parent commits.

I created a branch from next, and force merged tech/speed-up-ci-checks into it, resolving merge conflicts, but I'm very confused by the results, a lot of files changed I find puzzling.