@gaetanmaisse 2020-09-28
@ndelangen indeed that's weird, I still have `tech/speed-up-ci-checks` locally so I can try to update it with `next` and force push to see what happens.