@gaetanmaisse 2020-09-28
@ndelangen I rebased my branch and force-pushed it, the diff looks ok and it "only" has `+437 −360`