@yannbf 2020-09-28
The following changes [1](https://github.com/storybookjs/storybook/pull/12315/files#diff-9de053eff53b3da3ebb08e8a01781041R43) [2](https://github.com/storybookjs/storybook/pull/12315/files#diff-c61efba663a07d9aad178cf199022269R55) and [3](https://github.com/storybookjs/storybook/pull/12315/files#diff-6c085e0bdfbd605a8c503ba64854401fR62) will slightly impact new versions of `storybook/(aurelia,angular,addon-storyshots)`, should they be part of this PR as well?