@yannbf 2020-09-28
> @yannbf I think it's ok for `storybook/(angular,addon-storyshots)` as I just fixed some missing deps/peerDep. For `aurelia` I set a proper version as `dev` tag is used to publish daily build so I'm not sure we should use it on our side. I put them in this PR because I was focusing on being able to build the packages and run tests without cheating with the dependencies of the whole monorepo.
> 
> I'm not strongly opinionated so I can't revert it if you think I should. What's your opinion on that?

I'm not opinionated either, I was just thinking that this is a CI related PR which would/should not have production changes, and by touching these packages you might end up affecting their production code (as in a fix commit), even if it's just a peerDeps change. I'm not sure if it makes 100% sense, so don't feel blocked by this comment, I trust your senses! 👌