@tmeasday 2020-09-29
>> Replace CircleCI checkout step by a custom one to only checkout 1 commit and avoid downloading the whole monorepo and its history: 20s -> 1s for each step

> any idea what's up with all the chromatic changes?

These things are likely related 🤷‍♂️