@gaetanmaisse 2020-09-30
@tmeasday Thanks for pointing this out, I will take a look!

@yannbf If it's simple I will extract the few commits related to deps in another PR