@gaetanmaisse 2020-10-03
Thank you @tmeasday! I put back the "full" checkout check and chromatic is back to 🟢  🙇🏻 

@yannbf dep related commits were merged in another PR 🙂  

It looks good, latest runs (~ at the same time)
 - on next and another branch: `21min58`, `24min10`, `22min23`
 - on this branch: `15min43`, `14min47` 🎆 🎉 

@shilman @ndelangen I think this time we are good to go (except TeamCity 🤷🏻 )!