@shilman 2020-10-06
We can ignore TeamCity, they are not required